+++ src/lib.rs
@@ -2,12 +2,51 @@
 use core::{
     clone::Clone,
     fmt::{self, Debug, Formatter},
+    future::Future,
+    pin::Pin,
     sync::atomic::{AtomicBool, Ordering},
+    task::{Context, Poll, Waker},
 };
 use std::{
     sync::{Arc, Condvar, Mutex},
     thread,
 };
+
+struct FlowerState<SOME, OK>
+where
+    SOME: Clone + Send + Sync + 'static,
+    OK: Clone + Send + Sync + 'static,
+{
+    flowing: AtomicBool,
+    done: AtomicBool,
+    mtx: Mutex<(Option<SOME>, Option<OK>, Option<String>)>,
+    cvar: Condvar,
+    canceled: AtomicBool,
+}
+
+impl<SOME, OK> Debug for FlowerState<SOME, OK>
+where
+    SOME: Debug + Clone + Send + Sync + 'static,
+    OK: Debug + Clone + Send + Sync + 'static,
+{
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.debug_struct("FlowerState")
+            .field("flowing", &self.flowing)
+            .field("done", &self.done)
+            .field("mtx", &self.mtx)
+            .field("cvar", &self.cvar)
+            .field("canceled", &self.canceled)
+            .finish()
+    }
+}
+
+impl<SOME, OK> Drop for FlowerState<SOME, OK>
+where
+    SOME: Clone + Send + Sync + 'static,
+    OK: Clone + Send + Sync + 'static,
+{
+    fn drop(&mut self) {}
+}
 
 /// Flow loosely and gracefully.
 ///
@@ -31,6 +70,9 @@ use std::{
 ///                // Send current value through channel, will block the spawned thread
 ///                // until the option value successfully being polled in the main thread.
 ///                handle.send(i);
+///                // or handle.send_async(i).await; can be used from any async runtime,
+///                // it won't block the other async operations.  
+///                
 ///                // // Return error if the job is failure, for example:
 ///                // if i >= 3 {
 ///                //    return handle.err("Err".to_string());
@@ -44,7 +86,7 @@ use std::{
 ///    let mut exit = false;
 ///
 ///    loop {
-///        // Starting from version 1.x.x instead of polling the mutex over and over,
+///        // Starting from version 1.8.x instead of polling the mutex over and over,
 ///        // the fn will be activated automatically if the handle sending or return a value
 ///        // and will deactivate itself if the result value successfully received.
 ///        // Note: this fn is non-blocking (won't block the current thread).
@@ -76,8 +118,8 @@ where
     SOME: Clone + Send + Sync + 'static,
     OK: Clone + Send + Sync + 'static,
 {
-    state: Arc<(AtomicBool, Mutex<(Option<OK>, Option<String>)>)>,
-    flowing: Arc<(AtomicBool, Mutex<Option<SOME>>, Condvar, AtomicBool)>,
+    state: Arc<FlowerState<SOME, OK>>,
+    awaiting: Arc<(Mutex<Option<Waker>>, AtomicBool)>,
     id: usize,
 }
 
@@ -88,13 +130,14 @@ where
 {
     pub fn new(id: usize) -> Self {
         Self {
-            state: Arc::new((AtomicBool::new(false), Mutex::new((None, None)))),
-            flowing: Arc::new((
-                AtomicBool::new(false),
-                Mutex::new(None),
-                Condvar::new(),
-                AtomicBool::new(false),
-            )),
+            state: Arc::new(FlowerState {
+                flowing: AtomicBool::new(false),
+                done: AtomicBool::new(false),
+                mtx: Mutex::new((None, None, None)),
+                cvar: Condvar::new(),
+                canceled: AtomicBool::new(false),
+            }),
+            awaiting: Arc::new((Mutex::new(None), AtomicBool::new(false))),
             id,
         }
     }
@@ -106,9 +149,10 @@ where
 
     /// Get handle of the flower.
     pub fn handle(&self) -> FlowerHandle<SOME, OK> {
+        self.state.canceled.store(false, Ordering::Relaxed);
         FlowerHandle {
             state: Clone::clone(&self.state),
-            flowing: Clone::clone(&self.flowing),
+            awaiting: Clone::clone(&self.awaiting),
             id: self.id,
         }
     }
@@ -117,7 +161,12 @@ where
     ///
     /// will do noting if not explicitly configured.
     pub fn cancel(&self) {
-        self.flowing.3.store(true, Ordering::Relaxed);
+        self.state.canceled.store(true, Ordering::Relaxed);
+    }
+
+    /// Check if the flower is canceled
+    pub fn is_canceled(&self) -> bool {
+        self.state.canceled.load(Ordering::Relaxed)
     }
 
     /// Try receive the contained values of the flower
@@ -126,40 +175,41 @@ where
         c: Current,
         d: Done,
     ) {
-        if self.flowing.0.load(Ordering::Relaxed) {
-            if !self.state.0.load(Ordering::Relaxed) {
-                let value = match self.flowing.1.lock() {
-                    Ok(mut _result) => {
-                        let result = _result.clone();
-                        *_result = None;
-                        self.flowing.2.notify_one();
-                        if let Some(value) = result {
-                            value
-                        } else {
-                            return;
+        if self.state.flowing.load(Ordering::Relaxed) {
+            self.state.flowing.store(false, Ordering::Relaxed);
+            if !self.state.done.load(Ordering::Relaxed) {
+                let result = match self.state.mtx.lock() {
+                    Ok(mut result) => result.0.take(),
+                    _ => None,
+                };
+
+                if self.awaiting.1.load(Ordering::Relaxed) {
+                    self.awaiting.1.store(false, Ordering::Relaxed);
+                    if let Ok(mut mg_opt_waker) = self.awaiting.0.lock() {
+                        match mg_opt_waker.take() {
+                            Some(waker) => waker.wake(),
+                            _ => (),
                         }
                     }
-                    _ => return,
-                };
-                c(value);
+                } else {
+                    self.state.cvar.notify_one();
+                }
+
+                if let Some(value) = result {
+                    c(value);
+                }
             } else {
-                match self.state.1.lock() {
+                match self.state.mtx.lock() {
                     Ok(mut result) => {
-                        let (ok, error) = &mut *result;
-                        if let Some(val) = ok {
-                            let val = val.clone();
-                            *ok = None;
-                            d(Ok(val));
-                        } else if let Some(val) = error {
-                            let val = val.clone();
-                            *error = None;
-                            d(Err(val));
+                        let (_, ok, error) = &mut *result;
+                        if let Some(value) = ok.take() {
+                            d(Ok(value));
+                        } else if let Some(value) = error.take() {
+                            d(Err(value));
                         }
                     }
                     _ => (),
                 }
-                self.flowing.3.store(false, Ordering::Relaxed);
-                self.flowing.0.store(false, Ordering::Relaxed);
             }
         }
     }
@@ -173,7 +223,6 @@ where
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         f.debug_struct("Flower")
             .field("state", &self.state)
-            .field("flowing", &self.flowing)
             .field("id", &self.id)
             .finish()
     }
@@ -187,7 +236,7 @@ where
     fn clone(&self) -> Self {
         Self {
             state: Clone::clone(&self.state),
-            flowing: Clone::clone(&self.flowing),
+            awaiting: Clone::clone(&self.awaiting),
             id: self.id,
         }
     }
@@ -207,8 +256,8 @@ where
     SOME: Clone + Send + Sync + 'static,
     OK: Clone + Send + Sync + 'static,
 {
-    state: Arc<(AtomicBool, Mutex<(Option<OK>, Option<String>)>)>,
-    flowing: Arc<(AtomicBool, Mutex<Option<SOME>>, Condvar, AtomicBool)>,
+    state: Arc<FlowerState<SOME, OK>>,
+    awaiting: Arc<(Mutex<Option<Waker>>, AtomicBool)>,
     id: usize,
 }
 
@@ -224,29 +273,51 @@ where
 
     /// Check if the current flower should be canceled
     pub fn should_cancel(&self) -> bool {
-        self.flowing.3.load(Ordering::Relaxed)
+        self.state.canceled.load(Ordering::Relaxed)
     }
 
     /// Send current progress value
     pub fn send(&self, _value: SOME) {
-        let (_, mtx, cvar, _) = &*self.flowing;
-        if let Ok(mut mtx) = mtx.lock() {
-            *mtx = Some(_value);
-            self.state.0.store(false, Ordering::Relaxed);
-            self.flowing.0.store(true, Ordering::Relaxed);
-            let _ = cvar.wait(mtx);
+        if let Ok(mut mtx) = self.state.mtx.lock() {
+            mtx.0 = Some(_value);
+            self.state.done.store(false, Ordering::Relaxed);
+            self.state.flowing.store(true, Ordering::Relaxed);
+            self.awaiting.1.store(false, Ordering::Relaxed);
+            let _ = self.state.cvar.wait(mtx);
+        }
+    }
+
+    /// Send current progress value asynchronously.
+    pub async fn send_async(&self, _value: SOME) {
+        let mut pending = false;
+        match self.state.mtx.lock() {
+            Ok(mut mtx) => {
+                mtx.0 = Some(_value);
+                self.state.done.store(false, Ordering::Relaxed);
+                self.state.flowing.store(true, Ordering::Relaxed);
+                self.awaiting.1.store(true, Ordering::Relaxed);
+                pending = true;
+            }
+            _ => (),
+        };
+
+        if pending {
+            SendAsync {
+                awaiting: self.awaiting.clone(),
+            }
+            .await;
         }
     }
 
     /// Contains the success value for the result.
     pub fn ok(&self, _value: OK) {
-        match self.state.1.lock() {
+        match self.state.mtx.lock() {
             Ok(mut result) => {
-                let (ok, error) = &mut *result;
+                let (_, ok, error) = &mut *result;
                 *ok = Some(_value);
                 *error = None;
-                self.state.0.store(true, Ordering::Relaxed);
-                self.flowing.0.store(true, Ordering::Relaxed);
+                self.state.done.store(true, Ordering::Relaxed);
+                self.state.flowing.store(true, Ordering::Relaxed);
             }
             _ => (),
         }
@@ -254,15 +325,37 @@ where
 
     /// Contains the error value for the result.
     pub fn err(&self, _value: String) {
-        match self.state.1.lock() {
+        match self.state.mtx.lock() {
             Ok(mut result) => {
-                let (ok, error) = &mut *result;
+                let (_, ok, error) = &mut *result;
                 *error = Some(_value);
                 *ok = None;
-                self.state.0.store(true, Ordering::Relaxed);
-                self.flowing.0.store(true, Ordering::Relaxed);
+                self.state.done.store(true, Ordering::Relaxed);
+                self.state.flowing.store(true, Ordering::Relaxed);
             }
             _ => (),
+        }
+    }
+}
+
+struct SendAsync {
+    awaiting: Arc<(Mutex<Option<Waker>>, AtomicBool)>,
+}
+
+impl Future for SendAsync {
+    type Output = ();
+    fn poll(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Self::Output> {
+        match self.awaiting.0.lock() {
+            Ok(mut waker) => {
+                if !self.awaiting.1.load(Ordering::Relaxed) {
+                    Poll::Ready(())
+                } else {
+                    *waker = Some(_cx.waker().clone());
+                    Poll::Pending
+                }
+            }
+            // Rrevent blocking if something not ok with rust std mutex.
+            _ => Poll::Ready(()),
         }
     }
 }
@@ -275,7 +368,7 @@ where
     fn clone(&self) -> Self {
         Self {
             state: Clone::clone(&self.state),
-            flowing: Clone::clone(&self.flowing),
+            awaiting: Clone::clone(&self.awaiting),
             id: self.id,
         }
     }
@@ -288,7 +381,7 @@ where
 {
     fn drop(&mut self) {
         if thread::panicking() {
-            if !self.state.0.load(Ordering::Relaxed) {
+            if !self.state.done.load(Ordering::Relaxed) {
                 self.err(format!(
                     "the flower handle with id: {} error, the thread panicked maybe?",
                     self.id
@@ -306,10 +399,38 @@ where
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         f.debug_struct("FlowerHandle")
             .field("state", &self.state)
-            .field("flowing", &self.flowing)
             .field("id", &self.id)
             .finish()
     }
+}
+
+struct LeaperState<OK>
+where
+    OK: Clone + Send + Sync + 'static,
+{
+    leaping: AtomicBool,
+    mtx: Mutex<(Option<OK>, Option<String>)>,
+    canceled: AtomicBool,
+}
+
+impl<OK> Debug for LeaperState<OK>
+where
+    OK: Debug + Clone + Send + Sync + 'static,
+{
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.debug_struct("LeaperState")
+            .field("leaping", &self.leaping)
+            .field("mtx", &self.mtx)
+            .field("canceled", &self.canceled)
+            .finish()
+    }
+}
+
+impl<OK> Drop for LeaperState<OK>
+where
+    OK: Clone + Send + Sync + 'static,
+{
+    fn drop(&mut self) {}
 }
 
 /// (Flower without channel)
@@ -367,7 +488,7 @@ pub struct Leaper<OK>
 where
     OK: Clone + Send + Sync + 'static,
 {
-    state: Arc<(AtomicBool, Mutex<(Option<OK>, Option<String>)>, AtomicBool)>,
+    state: Arc<LeaperState<OK>>,
     id: usize,
 }
 
@@ -377,11 +498,11 @@ where
 {
     pub fn new(id: usize) -> Self {
         Self {
-            state: Arc::new((
-                AtomicBool::new(false),
-                Mutex::new((None, None)),
-                AtomicBool::new(false),
-            )),
+            state: Arc::new(LeaperState {
+                leaping: AtomicBool::new(false),
+                mtx: Mutex::new((None, None)),
+                canceled: AtomicBool::new(false),
+            }),
             id,
         }
     }
@@ -393,6 +514,7 @@ where
 
     /// Get handle of the leaper.
     pub fn handle(&self) -> LeaperHandle<OK> {
+        self.state.canceled.store(false, Ordering::Relaxed);
         LeaperHandle {
             state: Clone::clone(&self.state),
             id: self.id,
@@ -403,29 +525,29 @@ where
     ///
     /// will do noting if not explicitly configured.
     pub fn cancel(&self) {
-        self.state.2.store(true, Ordering::Relaxed);
+        self.state.canceled.store(true, Ordering::Relaxed);
+    }
+
+    /// Check if the leaper is canceled
+    pub fn is_canceled(&self) -> bool {
+        self.state.canceled.load(Ordering::Relaxed)
     }
 
     /// Try catch the value (result) of the leaper
     pub fn try_catch<Done: FnOnce(Result<OK, String>) -> ()>(&self, d: Done) {
-        if self.state.0.load(Ordering::Relaxed) {
-            match self.state.1.lock() {
+        if self.state.leaping.load(Ordering::Relaxed) {
+            self.state.leaping.store(false, Ordering::Relaxed);
+            match self.state.mtx.lock() {
                 Ok(mut result) => {
                     let (ok, error) = &mut *result;
-                    if let Some(val) = ok {
-                        let val = val.clone();
-                        *ok = None;
-                        d(Ok(val));
-                    } else if let Some(val) = error {
-                        let val = val.clone();
-                        *error = None;
-                        d(Err(val));
+                    if let Some(value) = ok.take() {
+                        d(Ok(value));
+                    } else if let Some(value) = error.take() {
+                        d(Err(value));
                     }
                 }
                 _ => (),
             }
-            self.state.0.store(false, Ordering::Relaxed);
-            self.state.2.store(false, Ordering::Relaxed);
         }
     }
 }
@@ -466,7 +588,7 @@ pub struct LeaperHandle<OK>
 where
     OK: Clone + Send + Sync + 'static,
 {
-    state: Arc<(AtomicBool, Mutex<(Option<OK>, Option<String>)>, AtomicBool)>,
+    state: Arc<LeaperState<OK>>,
     id: usize,
 }
 
@@ -481,17 +603,17 @@ where
 
     /// Check if the current leaper should be canceled
     pub fn should_cancel(&self) -> bool {
-        self.state.2.load(Ordering::Relaxed)
+        self.state.canceled.load(Ordering::Relaxed)
     }
 
     /// Contains the success value for the result.
     pub fn ok(&self, _value: OK) {
-        match self.state.1.lock() {
+        match self.state.mtx.lock() {
             Ok(mut result) => {
                 let (ok, error) = &mut *result;
                 *ok = Some(_value);
                 *error = None;
-                self.state.0.store(true, Ordering::Relaxed);
+                self.state.leaping.store(true, Ordering::Relaxed);
             }
             _ => (),
         }
@@ -499,12 +621,12 @@ where
 
     /// Contains the error value for the result.
     pub fn err(&self, _value: String) {
-        match self.state.1.lock() {
+        match self.state.mtx.lock() {
             Ok(mut result) => {
                 let (ok, error) = &mut *result;
                 *error = Some(_value);
                 *ok = None;
-                self.state.0.store(true, Ordering::Relaxed);
+                self.state.leaping.store(true, Ordering::Relaxed);
             }
             _ => (),
         }
@@ -529,7 +651,7 @@ where
 {
     fn drop(&mut self) {
         if thread::panicking() {
-            if !self.state.0.load(Ordering::Relaxed) {
+            if !self.state.leaping.load(Ordering::Relaxed) {
                 self.err(format!(
                     "the leaper handle with id: {} error, the thread panicked maybe?",
                     self.id